DEV: Big changes to allow shared library build
