+++ Code/CMakeLists.txt
@@ -84,7 +84,7 @@ endif()
 unset(SV_USE_FORTRAN)
 unset(SV_USE_MPI)
 unset(SV_USE_INTEL)
-unset(SV_USE_THEEDSOLVER_SVLS)
+unset(SV_THREEDSVOLER_USE_SVLS)
 
 #-----------------------------------------------------------------------------
 # Set EXE names for the executables
@@ -127,8 +127,14 @@ include(SetupSimVascularExternalsDefaults)
 
 #-----------------------------------------------------------------------------
 # Basic Options (CMAKE Defaults)
-option(BUILD_SHARED_LIBS "Build ${PROJECT_NAME} as shared libraries." OFF)
-mark_as_superbuild(VARS BUILD_SHARED_LIBS ALL_PROJECTS)
+option(SV_BUILD_SHARED_LIBS "Build ${PROJECT_NAME} as shared libraries." OFF)
+mark_as_superbuild(VARS SV_BUILD_SHARED_LIBS ALL_PROJECTS)
+set(SV_LIBRARY_TYPE "STATIC" CACHE STRING "Options are STATIC or SHARED" FORCE)
+set_property(CACHE SV_LIBRARY_TYPE PROPERTY STRINGS STATIC SHARED)
+mark_as_superbuild(SV_LIBRARY_TYPE)
+if(SV_BUILD_SHARED_LIBS)
+	set(SV_LIBRARY_TYPE "SHARED" CACHE STRING "Shared cache" FORCE)
+endif()
 
 find_package(Doxygen)
 if(DOXYGEN_FOUND)
@@ -154,7 +160,7 @@ endif(DOXYGEN_FOUND)
 #	include_directories(${FREETYPE_INCLUDE_DIR_ft2build})
 #endif()
 
-if(NOT BUILD_SHARED_LIBS)
+if(NOT SV_BUILD_SHARED_LIBS)
 	set(SV_INSTALL_HEADERS OFF)
 	set(SV_INSTALL_EXTERNALS OFF)
 	set(SV_INSTALL_LIBS OFF)
@@ -205,8 +211,8 @@ mark_as_superbuild(SV_USE_GTS)
 # Plugins
 option(SV_USE_ITK "Enable ITK Plugin" ON)
 mark_as_superbuild(SV_USE_ITK)
-option(ITK_USE_SHARED_LIBRARIES "Buiild vtk libraries as shared libs" OFF)
-mark_as_superbuild(ITK_USE_SHARED_LIBRARIES)
+option(ITK_SHARED_LIBRARIES "Buiild vtk libraries as shared libs" OFF)
+mark_as_superbuild(ITK_SHARED_LIBRARIES)
 
 option(SV_USE_VMTK "Enable VMTK Plugin" ON)
 mark_as_superbuild(SV_USE_VMTK)
@@ -223,13 +229,13 @@ if(SV_USE_TET_ADAPTOR)
 	set(SV_USE_THREEDSOLVER ON)
 endif()
 
-option(VTK_USE_SHARED_LIBRARIES "Buiild vtk libraries as shared libs" OFF)
-mark_as_superbuild(VTK_USE_SHARED_LIBRARIES)
+option(VTK_SHARED_LIBRARIES "Buiild vtk libraries as shared libs" OFF)
+mark_as_superbuild(VTK_SHARED_LIBRARIES)
 
 option(SV_USE_PYTHON "Use Tcl Python" OFF)
 mark_as_superbuild(SV_USE_PYTHON)
 if(SV_USE_PYTHON)
-	set(VTK_USE_SHARED_LIBRARIES "ON" CACHE BOOL "Initial cache" FORCE)
+	set(VTK_SHARED_LIBRARIES "ON" CACHE BOOL "Initial cache" FORCE)
 endif()
 
 #-----------------------------------------------------------------------------
@@ -248,10 +254,10 @@ endif()
 option(SV_USE_OPENCASCADE "OPENCASCADE" OFF)
 mark_as_superbuild(SV_USE_OPENCASCADE)
 if(SV_USE_OPENCASCADE)
-    set(VTK_USE_SHARED_LIBRARIES "ON" CACHE BOOL "Initial cache" FORCE)
+    set(VTK_SHARED_LIBRARIES "ON" CACHE BOOL "Initial cache" FORCE)
 endif()
 
-if(VTK_USE_SHARED_LIBRARIES)
+if(VTK_SHARED_LIBRARIES)
   if(UNIX AND NOT APPLE)
   	set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=gnu++11")
   endif()
@@ -303,17 +309,16 @@ if(SV_USE_THREEDSOLVER)
 	option(SV_THREEDSOLVER_USE_CLOSEDLOOP "" ON)
 	option(SV_THREEDSOLVER_USE_VARWALL "" ON)
 	option(SV_THREEDSOLVER_USE_VTK "" OFF)
+	mark_as_superbuild(SV_USE_THREEDSOLVER)
+	mark_as_superbuild(SV_THREEDSOLVER_USE_SOLVERIO)
+	mark_as_superbuild(SV_THREEDSOLVER_USE_SVPRE)
+	mark_as_superbuild(SV_THREEDSOLVER_USE_SVPOST)
+	mark_as_superbuild(SV_THREEDSOLVER_SOLVERIO_REDIRECT)
+	mark_as_superbuild(SV_THREEDSOLVER_USE_CORONARY)
+	mark_as_superbuild(SV_THREEDSOLVER_USE_CLOSEDLOOP)
+	mark_as_superbuild(SV_THREEDSOLVER_USE_VARWALL)
+	mark_as_superbuild(SV_THREEDSOLVER_USE_VTK)
 endif()
-mark_as_superbuild(SV_USE_THREEDSOLVER)
-mark_as_superbuild(SV_THREEDSOLVER_USE_SOLVERIO)
-mark_as_superbuild(SV_THREEDSOLVER_USE_SVPRE)
-mark_as_superbuild(SV_THREEDSOLVER_USE_SVPOST)
-mark_as_superbuild(SV_THREEDSOLVER_SOLVERIO_REDIRECT)
-mark_as_superbuild(SV_THREEDSOLVER_USE_CORONARY)
-mark_as_superbuild(SV_THREEDSOLVER_USE_CLOSEDLOOP)
-mark_as_superbuild(SV_THREEDSOLVER_USE_VARWALL)
-mark_as_superbuild(SV_THREEDSOLVER_USE_VTK)
-
 
 #-----------------------------------------------------------------------------
 # Linear Solver Options: SVLS
@@ -357,7 +362,7 @@ if(CMAKE_COMPILER_IS_GNUCXX)
 	set(GLOBAL_DEFINES "${GLOBAL_DEFINES} -DGCC")
 endif()
 
-if(NOT BUILD_SHARED_LIBS)
+if(NOT SV_BUILD_SHARED_LIBS)
 	set(SV_STATIC_BUILD "1")
 	set(GLOBAL_DEFINES "${GLOBAL_DEFINES} -DCV_STATIC_LINK -DSV_STATIC_BUILD")
 elseif()
@@ -586,7 +591,7 @@ endif()
 #-----------------------------------------------------------------------------
 # Plugin LevelSets
 if(SV_USE_ITK)
-	if(VTK_USE_SHARED_LIBRARIES)
+	if(ITK_SHARED_LIBRARIES)
 		set(SV_EXTERNAL_SHARED_LIBS ${SV_EXTERNAL_SHARED_LIBS} ITK)
 		set(SV_INSTALL_HEADERS ON)
 		set(SV_INSTALL_EXTERNALS ON)
@@ -601,12 +606,18 @@ endif()
 if(SV_USE_MESHSIM OR SV_USE_MESHSIM_DISCRETE_SOLIDMODEL)
 	set(SV_CORE_LIBRARY_DIRS ${SV_CORE_LIBRARY_DIRS} "Source/Mesh/MeshSimMeshObject")
 	if(SV_USE_MESHSIM)
-		option(MESHSIM_USE_SHARED_LIBRARIES "Build MeshSim as shared libraries" OFF)
-		mark_as_superbuild(MESHSIM_USE_SHARED_LIBRARIES)
+		option(SV_USE_MESHSIM_SHARED_LIBRARIES "Build MeshSim as shared libraries" OFF)
+		if(SV_BUILD_SHARED_LIBS)
+			set(SV_USE_MESHSIM_SHARED_LIBRARIES ON)
+		endif()
+		mark_as_superbuild(SV_USE_MESHSIM_SHARED_LIBRARIES)
 	endif()
 	if(SV_USE_MESHSIM_DISCRETE_SOLIDMODEL)
-		option(MESHSIM_DISCRETE_USE_SHARED_LIBRARIES "Build MeshSim Discrete Solid as shared libraries" OFF)
-		mark_as_superbuild(MESHSIM_DISCRETE_USE_SHARED_LIBRARIES)
+		option(SV_USE_MESHSIM_DISCRETE_SHARED_LIBRARIES "Build MeshSim Discrete Solid as shared libraries" OFF)
+		if(SV_BUILD_SHARED_LIBS)
+			set(SV_USE_MESHSIM_DISCRETE_SHARED_LIBRARIES ON)
+		endif()
+		mark_as_superbuild(SV_USE_MESHSIM_DISCRETE_SHARED_LIBRARIES)
 	endif()
 endif()
 
@@ -673,8 +684,11 @@ if(SV_USE_PARASOLID)
 			SET(USE_PARASOLID ON)
 			set(GLOBAL_DEFINES "${GLOBAL_DEFINES} -DUSE_PARASOLID")
 			set(SV_CORE_LIBRARY_DIRS ${SV_CORE_LIBRARY_DIRS} "Licensed/ParasolidSolidModel")
-			option(PARASOLID_USE_SHARED_LIBRARIES "Build Parasolid as shared libraries" OFF)
-			mark_as_superbuild(PARASOLID_USE_SHARED_LIBRARIES)
+			option(SV_USE_PARASOLID_SHARED_LIBRARIES "Build Parasolid as shared libraries" OFF)
+			if (SV_BUILD_SHARED_LIBS)
+				set(SV_USE_PARASOLID_SHARED_LIBRARIES ON)
+			endif()
+			mark_as_superbuild(SV_USE_PARASOLID_SHARED_LIBRARIES)
 		else()
 			message("Parasolid requires an extra license.")
 		endif()
@@ -904,7 +918,7 @@ endif()
 
 #-----------------------------------------------------------------------------
 # Set flags for shared libs
-if (PARASOLID_USE_SHARED_LIBRARIES OR MESHSIM_USE_SHARED_LIBRARIES)
+if (SV_USE_PARASOLID_SHARED_LIBRARIES OR SV_USE_MESHSIM_SHARED_LIBRARIES)
 	set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -fPIC")
 	set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -fPIC")
 endif()
@@ -1021,7 +1035,7 @@ if(${VTK_DIR} MATCHES "intel")
 	set(VTK_LIBRARIES ${VTK_LIBRARIES} ${INTELRUNTIME_LIBRARIES})
 endif()
 set(VTK_LIBRARIES ${VTK_LIBRARIES} vtkCommonCoreTCL)
-if(VTK_USE_SHARED_LIBRARIES)
+if(VTK_SHARED_LIBRARIES)
   set(GLOBAL_DEFINES "${GLOBAL_DEFINES} -DVTK_BUILD_SHARED_LIBS")
 endif()
 
@@ -1053,7 +1067,7 @@ if(SV_USE_PARASOLID)
 	if(PARASOLID_FOUND)
 		include_directories(${PARASOLID_INCLUDE_DIR})
 	endif()
-	if (PARASOLID_USE_SHARED_LIBRARIES)
+	if (SV_USE_PARASOLID_SHARED_LIBRARIES)
 		set(GLOBAL_DEFINES "${GLOBAL_DEFINES} -DUSE_PARASOLID_SHARED")
 	endif()
 endif()
@@ -1068,13 +1082,13 @@ if(SV_USE_MESHSIM)
 	if(MESHSIM_FOUND)
 		include_directories(${MESHSIM_INCLUDE_DIR})
 	endif()
-	if (MESHSIM_USE_SHARED_LIBRARIES)
+	if (SV_USE_MESHSIM_SHARED_LIBRARIES)
 		set(GLOBAL_DEFINES "${GLOBAL_DEFINES} -DUSE_MESHSIM_SHARED")
 	endif()
 endif()
 
 if (SV_USE_MESHSIM_DISCRETE_SOLIDMODEL)
-	if (MESHSIM_DISCRETE_USE_SHARED_LIBRARIES)
+	if (SV_USE_MESHSIM_DISCRETE_SHARED_LIBRARIES)
 		set(GLOBAL_DEFINES "${GLOBAL_DEFINES} -DUSE_DISCRETE_MODEL_SHARED")
 	endif()
 endif()
