+++ Code/Licensed/CMakeLists.txt
@@ -31,23 +31,22 @@
 set(SV_INCLUDE_DIRS_WORK ${SV_INCLUDE_DIRS})
 set(INDENTS "${INDENTS}")
 
-if(Licensed_ADAPTOR)
+if(SV_USE_MESHSIM_ADAPTOR)
 	set(SV_LICENSED_EXECUTABLE_DIRS ${SV_LICENSED_EXECUTABLE_DIRS} Adaptor)
 endif()
 
-if(Licensed_MESHSIM OR Licensed_MESHSIM_DISCRETE_SOLIDMODEL)
+if(SV_USE_MESHSIM OR SV_USE_MESHSIM_DISCRETE_SOLIDMODEL)
 	set(SV_LICENSED_LIBRARY_DIRS ${SV_LICENSED_LIBRARY_DIRS} MeshSimMeshObject)
 endif()
 
-if(NOT SV_EXCLUDE_SOLID_MODEL)
-	if(Licensed_PARASOLID)
-		set(SV_LICENSED_LIBRARY_DIRS ${SV_LICENSED_LIBRARY_DIRS} ParasolidSolidModel)
-	endif()
-
-	if(Licensed_MESHSIM_DISCRETE_SOLIDMODEL)
-		set(SV_LICENSED_LIBRARY_DIRS ${SV_LICENSED_LIBRARY_DIRS} MeshSimDiscreteSolidModel)
-	endif()
+if(SV_USE_PARASOLID)
+	set(SV_LICENSED_LIBRARY_DIRS ${SV_LICENSED_LIBRARY_DIRS} ParasolidSolidModel)
 endif()
+
+if(SV_USE_MESHSIM_DISCRETE_SOLIDMODEL)
+	set(SV_LICENSED_LIBRARY_DIRS ${SV_LICENSED_LIBRARY_DIRS} MeshSimDiscreteSolidModel)
+endif()
+
 if(SV_SUPERBUILD)
 	RETURN()
 	###### RETURN_SUPERBUILD ######
