+++ Code/Source/Common/Geometry/cv_geom_init.cxx
@@ -36,13 +36,16 @@
 #include <string.h>
 #include "cvRepository.h"
 #include "cvRepositoryData.h"
-#include "cvSolidModel.h"
 #include "cvPolyData.h"
 #include "cv_geom_init.h"
 #include "cv_sys_geom.h"
+#include "cvSolidModel.h"
+#include "cv_solid_init.h"
 #include "cv_integrate_surface.h"
 #include "cv_arg.h"
 #include "cv_misc_utils.h"
+#include "cv_vtk_utils.h"
+#include "cv_solid_init.h"
 
 // The following is needed for Windows
 #ifdef GetObject
@@ -246,6 +249,14 @@ int Geom_SetArrayForLocalOp_CellsCmd( ClientData clientData, Tcl_Interp *interp,
 
 int Geom_SetArrayForLocalOp_BlendCmd( ClientData clientData, Tcl_Interp *interp,
                            int argc, CONST84 char *argv[] );
+
+int Geom_All_UnionCmd( ClientData clientData, Tcl_Interp *interp,
+			   int argc, CONST84 char *argv[] );
+
+int Geom_Convert_NURBS_To_PolyCmd( ClientData clientData, Tcl_Interp *interp,
+			   int argc, CONST84 char *argv[] );
+
+
 #ifdef USE_GTS
 int Geom_Union_GTSCmd( ClientData clientData, Tcl_Interp *interp,
                            int argc, CONST84 char *argv[] );
@@ -420,6 +431,10 @@ int Geom_Init( Tcl_Interp *interp )
  Tcl_CreateCommand( interp, "geom_set_array_for_local_op_cells", Geom_SetArrayForLocalOp_CellsCmd,
 		     (ClientData)NULL, (Tcl_CmdDeleteProc *)NULL );
  Tcl_CreateCommand( interp, "geom_set_array_for_local_op_face_blend", Geom_SetArrayForLocalOp_BlendCmd,
+		     (ClientData)NULL, (Tcl_CmdDeleteProc *)NULL );
+  Tcl_CreateCommand( interp, "geom_all_union", Geom_All_UnionCmd,
+		     (ClientData)NULL, (Tcl_CmdDeleteProc *)NULL );
+  Tcl_CreateCommand( interp, "model_name_model_from_polydata_names", Geom_Convert_NURBS_To_PolyCmd,
 		     (ClientData)NULL, (Tcl_CmdDeleteProc *)NULL );
 #ifdef USE_GTS
   Tcl_CreateCommand( interp, "geom_union_gts", Geom_Union_GTSCmd,
@@ -1642,6 +1657,275 @@ int Geom_LocalBlendCmd( ClientData clientData, Tcl_Interp *interp,
 
   return TCL_OK;
 }
+
+// -------------
+// Geom_All_UnionCmd
+// -------------
+
+int Geom_All_UnionCmd( ClientData clientData, Tcl_Interp *interp,
+				 int argc, CONST84 char *argv[] )
+{
+  char *usage;
+  int numSrcs;
+  int interT;
+  ARG_List srcList;
+  char *dstName;
+  cvRepositoryData *src;
+  cvPolyData *dst;
+  RepositoryDataT type;
+  cvPolyData **srcs;
+  cvSolidModel *geom;
+  double tolerance = 1e-5;
+
+  int table_size = 4;
+  ARG_Entry arg_table[] = {
+    { "-srclist", LIST_Type, &srcList, NULL, REQUIRED, 0, { 0 } },
+    { "-intertype", INT_Type, &interT, NULL, REQUIRED, 0, { 0 } },
+    { "-result", STRING_Type, &dstName, NULL, REQUIRED, 0, { 0 } },
+    { "-tolerance", DOUBLE_Type, &tolerance, NULL, GDSC_OPTIONAL, 0, { 0 } },
+  };
+  usage = ARG_GenSyntaxStr( 1, argv, table_size, arg_table );
+  if ( argc == 1 ) {
+    Tcl_SetResult( interp, usage, TCL_VOLATILE );
+    return TCL_OK;
+  }
+  if ( ARG_ParseTclStr( interp, argc, argv, 1,
+			table_size, arg_table ) != TCL_OK ) {
+    Tcl_SetResult( interp, usage, TCL_VOLATILE );
+    return TCL_ERROR;
+  }
+
+  // Do work of command:
+  numSrcs = srcList.argc;
+
+  // Foreach src obj, check that it is in the repository and of the
+  // correct type (i.e. cvSolidModel).  Also build up the array of
+  // cvSolidModel*'s to pass to cvSolidModel::MakeLoftedSurf.
+
+  srcs = new cvPolyData * [numSrcs];
+
+  for (int i = 0; i < numSrcs; i++ ) {
+    src = gRepository->GetObject( srcList.argv[i] );
+    if ( src == NULL ) {
+      Tcl_AppendResult( interp, "couldn't find object ", srcList.argv[i],
+			(char *)NULL );
+      ARG_FreeListArgvs( table_size, arg_table );
+      delete [] srcs;
+      return TCL_ERROR;
+    }
+    type = src->GetType();
+    if ( type != POLY_DATA_T ) {
+      Tcl_AppendResult( interp, "object ", srcList.argv[i],
+			" not of type cvPolyData", (char *)NULL );
+      ARG_FreeListArgvs( table_size, arg_table );
+      delete [] srcs;
+      return TCL_ERROR;
+    }
+    srcs[i] = (cvPolyData *) src;
+  }
+
+  // We're done with the src object names:
+  ARG_FreeListArgvs( table_size, arg_table );
+
+  // Make sure the specified result object does not exist:
+  if ( gRepository->Exists( dstName ) ) {
+    Tcl_AppendResult( interp, "object ", dstName, " already exists",
+		      (char *)NULL );
+    delete [] srcs;
+    return TCL_ERROR;
+  }
+
+  // Instantiate the new solid:
+  geom = cvSolidModel::DefaultInstantiateSolidModel( interp );
+  if ( geom == NULL ) {
+    delete [] srcs;
+    return TCL_ERROR;
+  }
+
+  if ( sys_geom_all_union( srcs, numSrcs,interT,tolerance,(cvPolyData**)(&dst) )
+       != CV_OK ) {
+    Tcl_SetResult( interp, "poly manipulation error", TCL_STATIC );
+    delete dst;
+    delete [] srcs;
+    delete geom;
+    return TCL_ERROR;
+  }
+
+  vtkPolyData *dstPd;
+  dstPd = dst->GetVtkPolyData();
+  geom->SetVtkPolyDataObject(dstPd);
+  if ( !( gRepository->Register( dstName, geom ) ) ) {
+    Tcl_AppendResult( interp, "error registering obj ", dstName,
+		      " in repository", (char *)NULL );
+    delete geom;
+    delete [] srcs;
+    delete dst;
+    return TCL_ERROR;
+  }
+
+  // Make a new Tcl command:
+  Tcl_SetResult( interp, geom->GetName(), TCL_VOLATILE );
+  Tcl_CreateCommand( interp, Tcl_GetStringResult(interp), Solid_ObjectCmd,
+		     (ClientData)geom, DeleteSolid );
+
+  return TCL_OK;
+}
+
+// -------------
+// Geom_Convert_NURBS_To_PolyCmd
+// -------------
+
+int Geom_Convert_NURBS_To_PolyCmd( ClientData clientData, Tcl_Interp *interp,
+				 int argc, CONST84 char *argv[] )
+{
+  char *usage;
+  int numFaces;
+  int numIds;
+  int *allids;
+  int interT;
+  ARG_List faceList;
+  ARG_List idList;
+  char *srcName;
+  char *dstName;
+  cvRepositoryData *face;
+  cvPolyData *dst;
+  cvRepositoryData *model;
+  RepositoryDataT type;
+  cvPolyData **faces;
+  cvSolidModel *geom;
+
+  int table_size = 4;
+  ARG_Entry arg_table[] = {
+    { "-model", STRING_Type, &srcName, NULL, REQUIRED, 0, { 0 } },
+    { "-facelist", LIST_Type, &faceList, NULL, REQUIRED, 0, { 0 } },
+    { "-ids", LIST_Type, &idList, NULL, REQUIRED, 0, { 0 } },
+    { "-result", STRING_Type, &dstName, NULL, REQUIRED, 0, { 0 } },
+  };
+  usage = ARG_GenSyntaxStr( 1, argv, table_size, arg_table );
+  if ( argc == 1 ) {
+    Tcl_SetResult( interp, usage, TCL_VOLATILE );
+    return TCL_OK;
+  }
+  if ( ARG_ParseTclStr( interp, argc, argv, 1,
+			table_size, arg_table ) != TCL_OK ) {
+    Tcl_SetResult( interp, usage, TCL_VOLATILE );
+    return TCL_ERROR;
+  }
+
+  // Do work of command:
+  numFaces = faceList.argc;
+  numIds = idList.argc;
+
+  if (numFaces != numIds)
+  {
+      Tcl_AppendResult( interp, "Number of Ids must equal number of faces!\n");
+      ARG_FreeListArgvs( table_size, arg_table );
+      return TCL_ERROR;
+  }
+
+  // Foreach src obj, check that it is in the repository and of the
+  // correct type (i.e. cvSolidModel).  Also build up the array of
+  // cvSolidModel*'s to pass to cvSolidModel::MakeLoftedSurf.
+
+  faces = new cvPolyData * [numFaces];
+
+  for (int i = 0; i < numFaces; i++ ) {
+    face = gRepository->GetObject( faceList.argv[i] );
+    if ( face == NULL ) {
+      Tcl_AppendResult( interp, "couldn't find object ", faceList.argv[i],
+			(char *)NULL );
+      ARG_FreeListArgvs( table_size, arg_table );
+      delete [] faces;
+      return TCL_ERROR;
+    }
+    type = face->GetType();
+    if ( type != POLY_DATA_T ) {
+      Tcl_AppendResult( interp, "object ", faceList.argv[i],
+			" not of type cvPolyData", (char *)NULL );
+      ARG_FreeListArgvs( table_size, arg_table );
+      delete [] faces;
+      return TCL_ERROR;
+    }
+    faces[i] = (cvPolyData *) face;
+  }
+
+  // Create an array, and for each id insert it into the array
+  allids = new int[numIds];
+  if ( ARG_ParseTclListStatic( interp, idList, INT_Type, allids, idList.argc, &numIds )
+       != TCL_OK ) {
+    Tcl_SetResult( interp, usage, TCL_VOLATILE );
+    ARG_FreeListArgvs( table_size, arg_table );
+    return TCL_ERROR;
+  }
+
+  // Make sure the specified result object does not exist:
+  if ( gRepository->Exists( dstName ) ) {
+    Tcl_AppendResult( interp, "object ", dstName, " already exists",
+		      (char *)NULL );
+    delete [] faces;
+    delete [] allids;
+    return TCL_ERROR;
+  }
+
+  // Retrieve cvPolyData source:
+  model = gRepository->GetObject( srcName );
+  if ( model == NULL ) {
+    Tcl_AppendResult( interp, "couldn't find object ", srcName, (char *)NULL );
+    delete [] faces;
+    delete [] allids;
+    return TCL_ERROR;
+  }
+  type = model->GetType();
+  if ( type != POLY_DATA_T ) {
+    Tcl_AppendResult( interp, "object ", srcName, " not of type cvPolyData",
+		      (char *)NULL );
+    delete [] faces;
+    delete [] allids;
+    return TCL_ERROR;
+  }
+
+  // We're done with the src object names:
+  ARG_FreeListArgvs( table_size, arg_table );
+
+  // Instantiate the new solid:
+  geom = cvSolidModel::DefaultInstantiateSolidModel( interp );
+  if ( geom == NULL ) {
+    delete [] faces;
+    delete [] allids;
+    return TCL_ERROR;
+  }
+
+  if ( sys_geom_assign_ids_based_on_faces((cvPolyData *)model,faces,numFaces,allids,(cvPolyData**)(&dst) )
+       != CV_OK ) {
+    Tcl_SetResult( interp, "poly manipulation error", TCL_STATIC );
+    delete dst;
+    delete [] faces;
+    delete [] allids;
+    return TCL_ERROR;
+  }
+
+  delete [] faces;
+  delete [] allids;
+
+  vtkPolyData *dstPd;
+  dstPd = dst->GetVtkPolyData();
+  geom->SetVtkPolyDataObject(dstPd);
+  if ( !( gRepository->Register( dstName, geom ) ) ) {
+    Tcl_AppendResult( interp, "error registering obj ", dstName,
+		      " in repository", (char *)NULL );
+    delete dst;
+    return TCL_ERROR;
+  }
+
+  // Make a new Tcl command:
+  Tcl_SetResult( interp, geom->GetName(), TCL_VOLATILE );
+  Tcl_CreateCommand( interp, Tcl_GetStringResult(interp), Solid_ObjectCmd,
+		     (ClientData)geom, DeleteSolid );
+
+  return TCL_OK;
+}
+
+
 
 #ifdef USE_GTS
 // -------------
