+++ Code/Source/Common/Utils/cv_vtk_utils.cxx
@@ -1628,3 +1628,105 @@ int VtkUtils_ReversePtList( int num, double ptsIn[], double *ptsOut[] )
 
   return CV_OK;
 }
+
+// ----------------------
+// VtkUtils_PDCheckArrayName
+// ----------------------
+/** 
+ * @brief Function to check is array with name exists in cell or point data
+ * @param object this is the object to check if the array exists
+ * @param datatype this is point or cell. point =0,cell=1
+ * @param arrayname this is the name of the array to check
+ * @reutrn this returns 1 if the array exists and zero if it doesn't
+ * or the function does not return properly.
+ */
+
+int VtkUtils_PDCheckArrayName(vtkPolyData *object,int datatype,std::string arrayname )
+{
+  vtkIdType i;
+  int numArrays;
+  int exists =0;
+
+  if (datatype == 0)
+  {
+    numArrays = object->GetPointData()->GetNumberOfArrays();
+    for (i=0;i<numArrays;i++)
+    {
+      if (!strcmp(object->GetPointData()->GetArrayName(i),arrayname.c_str()))
+      {
+	exists =1;
+      }
+    }
+  }
+  else 
+  {
+    numArrays = object->GetCellData()->GetNumberOfArrays();
+    for (i=0;i<numArrays;i++)
+    {
+      if (!strcmp(object->GetCellData()->GetArrayName(i),arrayname.c_str()))
+      {
+	exists =1;
+      }
+    }
+  }
+
+  if (exists == 1)
+  {
+    return CV_OK;
+  }
+  else
+  {
+    return CV_ERROR;
+  }
+}
+
+// -------------------
+// VtkUtils_UGCheckArrayName
+// -------------------
+/** 
+ * @brief Function to check is array with name exists in cell or point data
+ * @param object this is the object to check if the array exists
+ * @param datatype this is point or cell. point =0,cell=1
+ * @param arrayname this is the name of the array to check
+ * @reutrn this returns 1 if the array exists and zero if it doesn't
+ * or the function does not return properly.
+ */
+
+int VtkUtils_UGCheckArrayName(vtkUnstructuredGrid *object,int datatype,std::string arrayname)
+{
+  vtkIdType i;
+  int numArrays;
+  int exists =0;
+
+  if (datatype == 0)
+  {
+    numArrays = object->GetPointData()->GetNumberOfArrays();
+    for (i=0;i<numArrays;i++)
+    {
+      if (!strcmp(object->GetPointData()->GetArrayName(i),arrayname.c_str()))
+      {
+	exists =1;
+      }
+    }
+  }
+  else 
+  {
+    numArrays = object->GetCellData()->GetNumberOfArrays();
+    for (i=0;i<numArrays;i++)
+    {
+      if (!strcmp(object->GetCellData()->GetArrayName(i),arrayname.c_str()))
+      {
+	exists =1;
+      }
+    }
+  }
+
+  if (exists == 1)
+  {
+    return CV_OK;
+  }
+  else
+  {
+    return CV_ERROR;
+  }
+}
