+++ Code/Source/Legacy/LevelSet/CMakeLists.txt
@@ -22,7 +22,7 @@ SET(HDRS cvLevelSet.h cvLevelSetNode.h
     cvLevelSetVelocityImage.h
     )
 
-add_library(${lib} ${CXXSRCS} cv_LsetCore_init.cxx cv_LsetV_init.cxx)
+add_library(${lib} ${SV_LIBRARY_TYPE} ${CXXSRCS} cv_LsetCore_init.cxx cv_LsetV_init.cxx)
 
 target_link_libraries(${lib} 
    ${ITK_LIBRARIES} ${VTK_LIBRARIES}
