+++ Code/Source/Model/OCCTSolidModel/CMakeLists.txt
@@ -34,11 +34,16 @@
 
 set(lib lib_simvascular_opencascade)
 LIST(APPEND CORELIBS ${lib})
+if(SV_USE_PYTHON)
+	include_directories(${PYTHON_INCLUDE_PATH}
+		    ${PYTHON_INCLUDE_DIR}
+		    ${PYTHON_INCLUDE_DIRS})
+endif()
 
 SET(CXXSRCS cvOCCTSolidModel.cxx cv_occtsolid_utils.cxx)
 SET(HDRS cvOCCTSolidModel.h cv_occt_init.h cv_occtsolid_utils.h)
 
-add_library(${lib} ${CXXSRCS} cv_occt_init.cxx)
+add_library(${lib} ${SV_LIBRARY_TYPE} ${CXXSRCS} cv_occt_init.cxx)
 
 target_link_libraries(${lib}
 			${VTK_LIBRARIES} ${TCL_LIBRARY} ${TK_LIBRARY}
@@ -48,6 +53,9 @@ target_link_libraries(${lib}
 			lib_simvascular_solid lib_simvascular_repository
 			lib_simvascular_globals lib_simvascular_utils lib_simvascular_sysgeom
             )
+if(SV_USE_PYTHON)
+   target_link_libraries(${lib} ${PYTHON_LIBRARIES})
+endif()
 
 if(SV_INSTALL_LIBS)
 	install(TARGETS ${lib}
