+++ Code/Source/Model/PolyDataSolidModel/cv_polydatasolid_utils.cxx
@@ -46,6 +46,7 @@
 #include <string.h>
 #include "cv_polydatasolid_utils.h"
 #include "cv_misc_utils.h"
+#include "cv_vtk_utils.h"
 #include "cv_sys_geom.h"
 
 #include "cv_globals.h"
@@ -110,7 +111,7 @@ int PlyDtaUtils_GetFaceIds( vtkPolyData *geom, int *v_num_faces, int **v_faces)
   *v_faces = NULL;
 
   boundaryScalars = vtkSmartPointer<vtkIntArray>::New();
-  if (PlyDtaUtils_PDCheckArrayName(geom,1,"ModelFaceID") != CV_OK)
+  if (VtkUtils_PDCheckArrayName(geom,1,"ModelFaceID") != CV_OK)
   {
     fprintf(stderr,"Array name 'ModelFaceID' does not exist. Regions must be identified");
     fprintf(stderr," and named 'ModelFaceID' prior to this function call\n");
@@ -424,7 +425,7 @@ int PlyDtaUtils_CombineFaces(vtkPolyData *geom,int *targetface,int *loseface )
   id1 = *targetface;
   id2 = *loseface;
 
-  if (PlyDtaUtils_PDCheckArrayName(geom,1,"ModelFaceID") != CV_OK)
+  if (VtkUtils_PDCheckArrayName(geom,1,"ModelFaceID") != CV_OK)
   {
     fprintf(stderr,"Array name 'ModelFaceID' does not exist. Regions must be identified \
 		    and named 'ModelFaceID' prior to this function call\n");
@@ -493,7 +494,7 @@ int PlyDtaUtils_DeleteRegion(vtkPolyData *geom,int *regionid)
   vtkSmartPointer<vtkIntArray> boundaryRegions =
     vtkSmartPointer<vtkIntArray>::New();
 
-  if (PlyDtaUtils_PDCheckArrayName(geom,1,"ModelFaceID") != CV_OK)
+  if (VtkUtils_PDCheckArrayName(geom,1,"ModelFaceID") != CV_OK)
   {
     fprintf(stderr,"Array name 'ModelFaceID' does not exist. Regions must be identified \
 		    and named 'ModelFaceID' prior to this function call\n");
@@ -516,105 +517,3 @@ int PlyDtaUtils_DeleteRegion(vtkPolyData *geom,int *regionid)
   return CV_OK;
 }
 
-
-// -------------------
-// PlyDtaUtils_CheckArrayName
-// -------------------
-/** 
- * @brief Function to check is array with name exists in cell or point data
- * @param object this is the object to check if the array exists
- * @param datatype this is point or cell. point =0,cell=1
- * @param arrayname this is the name of the array to check
- * @reutrn this returns 1 if the array exists and zero if it doesn't
- * or the function does not return properly.
- */
-
-int PlyDtaUtils_PDCheckArrayName(vtkPolyData *object,int datatype,std::string arrayname)
-{
-  vtkIdType i;
-  int numArrays;
-  int exists =0;
-
-  if (datatype == 0)
-  {
-    numArrays = object->GetPointData()->GetNumberOfArrays();
-    for (i=0;i<numArrays;i++)
-    {
-      if (!strcmp(object->GetPointData()->GetArrayName(i),arrayname.c_str()))
-      {
-	exists =1;
-      }
-    }
-  }
-  else 
-  {
-    numArrays = object->GetCellData()->GetNumberOfArrays();
-    for (i=0;i<numArrays;i++)
-    {
-      if (!strcmp(object->GetCellData()->GetArrayName(i),arrayname.c_str()))
-      {
-	exists =1;
-      }
-    }
-  }
-
-  if (exists == 1)
-  {
-    return CV_OK;
-  }
-  else
-  {
-    return CV_ERROR;
-  }
-}
-
-// -------------------
-// PlyDtaUtils_UGCheckArrayName
-// -------------------
-/** 
- * @brief Function to check is array with name exists in cell or point data
- * @param object this is the object to check if the array exists
- * @param datatype this is point or cell. point =0,cell=1
- * @param arrayname this is the name of the array to check
- * @reutrn this returns 1 if the array exists and zero if it doesn't
- * or the function does not return properly.
- */
-
-int PlyDtaUtils_UGCheckArrayName(vtkUnstructuredGrid *object,int datatype,std::string arrayname)
-{
-  vtkIdType i;
-  int numArrays;
-  int exists =0;
-
-  if (datatype == 0)
-  {
-    numArrays = object->GetPointData()->GetNumberOfArrays();
-    for (i=0;i<numArrays;i++)
-    {
-      if (!strcmp(object->GetPointData()->GetArrayName(i),arrayname.c_str()))
-      {
-	exists =1;
-      }
-    }
-  }
-  else 
-  {
-    numArrays = object->GetCellData()->GetNumberOfArrays();
-    for (i=0;i<numArrays;i++)
-    {
-      if (!strcmp(object->GetCellData()->GetArrayName(i),arrayname.c_str()))
-      {
-	exists =1;
-      }
-    }
-  }
-
-  if (exists == 1)
-  {
-    return CV_OK;
-  }
-  else
-  {
-    return CV_ERROR;
-  }
-}
