+++ Code/Source/PostProcessing/CMakeLists.txt
@@ -12,14 +12,14 @@ SET(HDRS    cvConvertVisFiles.h cvCalculateWallShearStress.h
             cvCalculateTKE.h
             )
 
-add_library(${lib} ${CXXSRCS})
+add_library(${lib} ${SV_LIBRARY_TYPE} ${CXXSRCS})
 
 target_link_libraries(lib_simvascular_post
 			${VTK_LIBRARIES} ${TCL_LIBRARY} 
 			${TK_LIBRARY} 
 			${GLOBAL_LIBRARIES}
 			${ZLIB_LIBRARY}
-			lib_simvascular_globals)
+			lib_simvascular_globals lib_simvascular_repository)
 
 if(SV_INSTALL_LIBS)
 	install(TARGETS ${lib}
