+++ Code/SuperBuild/ITK.cmake
@@ -71,7 +71,7 @@ if(NOT ${CMAKE_PROJECT_NAME}_USE_SYSTEM_${proj})
    -DCMAKE_BUILD_TYPE:STRING=${CMAKE_BUILD_TYPE}
    -DCMAKE_THREAD_LIBS:STRING=-lpthread
    -DBUILD_EXAMPLES:BOOL=OFF
-   -DBUILD_SHARED_LIBS:BOOL=${ITK_USE_SHARED_LIBRARIES}
+   -DBUILD_SHARED_LIBS:BOOL=${ITK_SHARED_LIBRARIES}
    -DBUILD_TESTING:BOOL=OFF
    -DITK_WRAP_PYTHON:BOOL=OFF
    -DITK_LEGACY_SILENT:BOOL=OFF
