+++ Code/SuperBuild/VTK.cmake
@@ -86,7 +86,7 @@ if(NOT ${CMAKE_PROJECT_NAME}_USE_SYSTEM_${proj})
    -DCMAKE_C_FLAGS:STRING=${CMAKE_C_FLAGS}
    -DCMAKE_BUILD_TYPE:STRING=${CMAKE_BUILD_TYPE}
    -DCMAKE_THREAD_LIBS:STRING=-lpthread
-   -DBUILD_SHARED_LIBS:BOOL=${VTK_USE_SHARED_LIBRARIES}
+   -DBUILD_SHARED_LIBS:BOOL=${VTK_SHARED_LIBRARIES}
    -DBUILD_TESTING:BOOL=OFF
    -DVTK_WRAP_TCL:BOOL=ON
    -DVTK_WRAP_PYTHON:BOOL=${SV_USE_PYTHON}
