+++ Tcl/SimVascular_2.0/Plugins/tclpython.tcl
@@ -32,7 +32,10 @@ proc startTclPython {} {
   #Initiate modules
   puts "Initiate the python modules from tcl"
   if [catch {$gPythonInterp exec {tcl.eval("solid_initPyMods")}} errmsg] {
-    puts "Could not initialize internal python modules: $errmsg"
+    puts "Could not initialize internal python solid modules: $errmsg"
+  }
+  if [catch {$gPythonInterp exec {tcl.eval("occt_initPyMods")}} errmsg] {
+    puts "Could not initialize internal python occt modules: $errmsg"
   }
   if [catch {$gPythonInterp exec {import os}} errmsg] {
     puts "No os module found: $errmsg"
@@ -107,6 +110,7 @@ proc testCreateSurf {} {
   solid_newObject -name myNewObj
 
   ##Python Interp
+  $gPythonInterp exec {import pyOCCT}
   $gPythonInterp exec {import numpy as np}
   $gPythonInterp exec {X = np.ndarray.tolist(np.array([[3.1,4.3,5.3],[1.1,6.1,8.5]]))}
   $gPythonInterp exec {Y = np.ndarray.tolist(np.array([[3.1,4.3,5.3],[1.1,6.1,8.5]]))}
@@ -116,5 +120,5 @@ proc testCreateSurf {} {
   $gPythonInterp exec {vK = np.ndarray.tolist(np.array([3.0,4.0,51.0,6.0,8.0]))}
   $gPythonInterp exec {uM = np.ndarray.tolist(np.array([3.0,4.0,51.0,6.0,8.0]))}
   $gPythonInterp exec {vM = np.ndarray.tolist(np.array([3.0,4.0,51.0,6.0,8.0]))}
-  $gPythonInterp exec {pySolid.convertListsToOCCT("myNewObj",X,Y,Z,uK,vK,uM,vM,2,2)}
+  $gPythonInterp exec {pyOCCT.convertListsToOCCT("myNewObj",X,Y,Z,uK,vK,uM,vM,2,2)}
 }
